django-udemy-clone-master/migration_requirements.txt
mysql-connector-python==8.0.33
Django==5.1.7
mysqlclient==2.2.7 